+++ Logic/src/Score.cpp
@@ -30,7 +30,7 @@ logic::Score::Score() {
 
 void logic::Score::onNotify(const Subject& entity, Event& e) {
     if (e == Event::CoinCollected) {
-        double standard_increase = 10;
+        float standard_increase = 10;
         auto stopwatch = Stopwatch::getInstance();
         auto now = stopwatch->get_now();
 
@@ -39,7 +39,7 @@ void logic::Score::onNotify(const Subject& entity, Event& e) {
             first_coin_collected = true;
             score_ += standard_increase;
         } else {
-            double time_between = stopwatch->get_time_between(now, previous_coin_time);
+            float time_between = stopwatch->get_time_between(now, previous_coin_time);
             score_ += std::lround(standard_increase * (1 / (time_between / 1000)));
         }
         previous_coin_time = now;
